refactor(RF-001): extrai construção do status para módulo buildStatus (manter comportamento) [BLUE]
