+++ pages/api/v1/status/index.js
@@ -1,29 +1,15 @@
 import database from "infra/database.js";
-import fs from "fs";
-import path from "path";
-
-// Lê name e version do package.json
-const pkg = JSON.parse(
-  fs.readFileSync(path.join(process.cwd(), "package.json"), "utf-8"),
-);
+import { buildStatus } from "../../../../models/status/buildStatus.js";
 
 async function status(request, response) {
   try {
     const updatedAt = new Date().toISOString();
 
-    const queryVersionDataBaseResult = await database.query(
-      "SHOW server_version;",
-    );
-    const queryVersionDataBaseValue = parseInt(
-      queryVersionDataBaseResult.rows[0].server_version,
-    );
+    const queryVersionDataBaseResult = await database.query("SHOW server_version;");
+    const queryVersionDataBaseValue = parseInt(queryVersionDataBaseResult.rows[0].server_version);
 
-    const maxConnectionsDataBaseResult = await database.query(
-      "SHOW max_connections;",
-    );
-    const maxConnectionsDataBaseValue = parseInt(
-      maxConnectionsDataBaseResult.rows[0].max_connections,
-    );
+    const maxConnectionsDataBaseResult = await database.query("SHOW max_connections;");
+    const maxConnectionsDataBaseValue = parseInt(maxConnectionsDataBaseResult.rows[0].max_connections);
 
     const databaseName = process.env.POSTGRES_DB;
 
@@ -31,10 +17,9 @@ async function status(request, response) {
       text: "SELECT count(*)::int as count_connections FROM pg_stat_activity where datname = $1",
       values: [databaseName],
     });
-    const rountConnectionsDataBaseValue =
-      rountConnectionsDataBaseResult.rows[0].count_connections;
+    const rountConnectionsDataBaseValue = rountConnectionsDataBaseResult.rows[0].count_connections;
 
-    response.status(200).json({
+    const base = {
       updated_at: updatedAt,
       dependencies: {
         database: {
@@ -43,15 +28,9 @@ async function status(request, response) {
           count_connections: rountConnectionsDataBaseValue,
         },
       },
-      // >>> RF-001
-      application: {
-        name: pkg.name,
-        version: pkg.version,
-        uptime_seconds: Math.floor(process.uptime()),
-        environment: process.env.NODE_ENV || "development",
-      },
-      // <<< RF-001
-    });
+    };
+
+    response.status(200).json(buildStatus(base));
   } catch (error) {
     response.status(500).json({
       status: "error",
